requirements.txt

# Core dependencies
pyyaml>=6.0
jinja2>=3.0
click>=8.0
colorama>=0.4
pathlib>=1.0
Pillow>=9.0  # For screenshot generation
selenium>=4.0  # For automated VS Code screenshots
webdriver-manager>=3.8  # For managing browser drivers
wl_ai_manager>=0.1
wl_config_manager>=0.1
python-dotenv>=0.19
requests>=2.28
tqdm>=4.65  # Progress bars
termcolor>=2.3  # Colored terminal output
jsonschema>=4.17  # For validating theme schemas